refactor: update translation and enhance BelongsToMultiselectInput component
