+++ web_application/locales/de/general.json
@@ -12,7 +12,7 @@
     "made_with": "Digitales Produkt von",
     "nothing_found": "Keine Treffer",
     "privacy": "Datenschutz",
-    "search_select": "Suchen und wählen",
+    "search_select": "Suchen und hinzufügen",
     "select": "Wählen",
 
     "title": "Titel",
