+++ web_application/locales/en/general.json
@@ -12,7 +12,7 @@
     "made_with": "Digital product by",
     "nothing_found": "Nothing found",
     "privacy": "Privacy policy",
-    "search_select": "Search and select",
+    "search_select": "Search and add",
     "select": "Select",
 
     "title": "Title",
